Use page data property select when adding position placeholder
